got the desktop stuff actually working
fixed bugs with rust generation. there was a bunch of stuff but it basically boiled down to the parameterization loop not being initialized properly, some incorrect handling of vector as a result of incorrect translation from C++ to Rust, and a couple other random issues here and there. Also added a debug print for when trajectory is called and some debug prints for errors.